Feature: -Preset Parameter
Added a -Preset parameter that has several easy to use presets ready to go.

JustinGrote CI: Add Verbose to IB Tests
ae8f278
JustinGrote Refactor: Update Module Import Method
f70fd32
JustinGrote Test: Remove Version from Test Dependencies
f5e4181
JustinGrote Add Preset Logic and Multiple Output Types
44f7a0e
JustinGrote Test: Fixes to improve performance and accuracy
de860c8
JustinGrote Bugfix: Invoke-NmapExe returned multiline input
cc80a22
JustinGrote Rearrange Presets
c84d0c1
JustinGrote Refactor: Output Forms
bc2b752
JustinGrote Move to Private Function
ef55008
JustinGrote Test: Add Verbose message about what NMAP command is being executed
6051b11
JustinGrote Feature: Add ThrowUser command

+++ PoshNmap/Public/Invoke-Nmap.ps1
@@ -1,31 +1,66 @@
+Register-ArgumentCompleter -CommandName "Invoke-Nmap" -ParameterName "Preset" -ScriptBlock {(Get-NmapPresetArguments).keys}
+
 function Invoke-Nmap {
     <#
     .SYNOPSIS
         Runs the NMAP command and then formats the output.
-    .EXAMPLE
-        Invoke-Nmap www.google.com
-        Runs an NMAP scan of www.google.com and provides the results as powershell objects
     .INPUTS
         [String[]]$ComputerName
     .OUTPUTS
-        [PoshNMAP.NMAPResult]
+        Depends on -OutFormat parameter
+    .EXAMPLE
+        Invoke-Nmap scanme.nmap.org
+        Runs an NMAP scan with the Quick scan preset and provides the result as a formatted Powershell Object
+    .EXAMPLE
+        Invoke-Nmap scanme.nmap.org "-t4 -p 80,443"
+        This is similar to running nmap "bare" but enjoy the format processing of invoke-nmap
+
     #>
 
-    [CmdletBinding(DefaultParameterSetName="default")]
+    [CmdletBinding(DefaultParameterSetName="preset")]
     param (
-        #A list of nmap host specifications
-        [String[]][Parameter(Position=0,ParameterSetName="Default",ValueFromPipeline)]$computerName = "localhost",
+        #A list of nmap host specifications. Defaults to localhost
+        [Parameter(
+            Position=0,
+            ValueFromPipeline
+        )]
+        [String[]]
+        $computerName = "localhost",
+
+        #Specify raw argument parameters to nmap
+        [String[]]
+        [Parameter(
+            ParameterSetName="custom",
+            Mandatory,
+            Position=1
+        )]
+        $ArgumentList,
+
+        [String]
+        [Parameter(
+            ParameterSetName="preset",
+            ValueFromPipeline
+        )]
+        [String]$Preset = "Quick",
+
+        #Choose which format for the output (XML, JSON, HashTable, PSObject, or Raw). Default is PSObject
+        [ValidateSet('Raw','PSObject','XML','JSON','Hashtable')]
+        [String]$OutFormat = 'PSObject',
+
         #A list of SNMP communities to scan. Defaults to public and private
-        [String[]][Parameter(ParameterSetName="Default")]$snmpCommunityList = @("private","public"),
-        #Specify this if you want the raw (non-Powershell-formatted) NMAP Output
-        [Switch]$Raw,
-        #Override the default nmap parameters
-        $ArgumentList = @(
-            "--open",
-            "-T4",
-            "-F"
-        )
+        [String[]]
+        [Parameter()]
+        $snmpCommunityList = @("private","public")
     )
+
+    if ($Preset) {
+        $nmapPresetArgumentNames = (Get-NmapPresetArguments).keys
+        if ($Preset -notin $nmapPresetArgumentNames) {
+            throwUser New-Object ArgumentException -ArgumentList "Invoke-Nmap: Value $Preset is not a valid choice. Please choose one of: $($nmapPresetArgumentNames -join ', ')","Preset"
+        } else {
+            $ArgumentList = Get-NmapPresetArguments $Preset
+        }
+    }
 
     if ($snmpCommunityList) {
         $snmpCommunityFile = [io.path]::GetTempFileName()
@@ -34,17 +69,32 @@ function Invoke-Nmap {
     }
     $nmapexe = 'nmap.exe'
 
-    if (-not $Raw) {
-        $ArgumentList += "-oX","-"
+    if ($OutFormat -eq 'Raw') {
+        Invoke-NmapExe $nmapExe $argumentList $computerName
+        exit $LASTEXITCODE
     }
 
-    $nmapresult = Invoke-NmapExe $nmapExe $argumentList $computerName
-
-    if (-not $nmapResult) {throw "NMAP did not produce any output. Please review any errors that are present above this warning."}
-    if ($Raw) {
-        $nmapResult
-    } else {
-        [xml]$nmapresult | ConvertFrom-NmapXML -Format HashTable | % nmaprun
+    $ArgumentList += "-oX","-"
+    try {
+        [String]$nmapresult = Invoke-NmapExe $nmapExe $argumentList $computerName
+    } finally {
+        if (Test-Path $snmpCommunityFile) {Remove-Item $snmpCommunityFile -Force -ErrorAction SilentlyContinue}
     }
-    Remove-Item $snmpCommunityFile -Force -ErrorAction SilentlyContinue
+
+
+    if (-not $nmapResult) {throwUser "NMAP did not produce any output. Please review any errors that are present above this warning."}
+    switch ($OutFormat) {
+        'XML' {
+            $nmapResult
+        }
+        'JSON' {
+            $nmapResult | ConvertFrom-NmapXML
+        }
+        'PSObject' {
+            $nmapResult | ConvertFrom-NmapXML -OutFormat PSObject
+        }
+        'HashTable' {
+            $nmapResult | ConvertFrom-NmapXML -OutFormat HashTable
+        }
+    }
 }
